Attempt to fix CI error by moving files to a different folder
