+++ src/pages/Search/components/SearchPage.jsx
@@ -4,7 +4,7 @@ import { useLocation, useNavigate } from "react-router-dom";
 import axios from "axios";
 import throttle from "lodash.throttle";
 
-import SearchBar from "src/components/SearchBar/SearchBar";
+import SearchBar from "src/components/RandomFilename/SearchBar";
 import AdvancedSearchMenu from "src/pages/Search/components/AdvancedSearch/AdvancedSearch";
 import SearchCard from "src/pages/Search/components/SearchCard/SearchCard.jsx";
 import { capitalize } from "src/utils/common";
