+++ tests/UnitTest/Spoonacular/SearchByIngridients/SearchBar.test.jsx
@@ -1,7 +1,7 @@
 import { describe, test, expect } from "vitest";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import SearchBar from "src/components/SearchBar/SearchBar";
+import SearchBar from "src/components/RandomFilename/SearchBar";
 
 describe("SearchBar", async () => {
   test("renders input field and button", () => {
